Issues:
-
https://github.com/threesigmaxyz/mzero-labs-8-1-2024-issues-external/issues/9
& ChainSecurity CS-MZEROCORE-027
- fixed by
https://github.com/MZero-Labs/ttg/pull/231/commits/b97b196bc04bc8104b9bf07d1dd5570fb71f18ae
- ChainSecurity CS-MZEROCORE-021
- fixed by
https://github.com/MZero-Labs/ttg/pull/231/commits/ae89c920bee3e9331c7d42b7da324752226bc4e8

+++ test/fuzz/StandardGovernorFuzz.t.sol
@@ -75,7 +75,7 @@ contract StandardGovernorTests is TestUtils {
     }
 
     /* ============ castVotes ============ */
-    function testFuzz_castVotes(uint numberOfProposals) external {
+    function testFuzz_castVotes(uint256 numberOfProposals) external {
         numberOfProposals = uint8(bound(numberOfProposals, 1, type(uint8).max));
 
         uint256[] memory proposalIds_ = new uint256[](numberOfProposals);
